+++ packages/react-graphql/CHANGELOG.md
@@ -9,7 +9,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Added
 
-- Adds SSR trackable link to support GraphQL Self-serialization ([810](https://github.com/Shopify/quilt/pull/810))
+- Adds SSR extractable link to support GraphQL Self-serialization ([810](https://github.com/Shopify/quilt/pull/810))
 
 ## [4.0.2] - 2019-07-05